backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
httpx==0.27.0
pytest==8.3.2
ruff==0.5.7
black==24.8.0
mypy==1.11.1
loguru==0.7.2
python-jose[cryptography]==3.3.0

